Fixed issue on sending data over
We had created an array of particles and stored it into a pointer on our end after zeroing the spot. This meant that although the first particle existed we had only zeros in the actual data spots after. We now just set a new particle there when emitting a new particle